pyproject.toml
[project]
name = "alrocar-weather-service"
version = "0.2.2"
description = "A MCP server project to integrate OpenWeather API in Claude"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
 "httpx>=0.27.2",
 "mcp>=1.0.0",
 "python-dotenv>=1.0.1",
]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
alrocar-weather-service = "alrocar_weather_service:main"

[project.optional-dependencies]
dev = [
    "black>=23.12.1",
    "pyproject-toml>=0.0.10",
]

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
